// types/mailchimp-analytics.ts
export interface MailchimpEngagementData {
  campaignId: string
  campaignTitle: string
  sendTime: string
  opens: number
  uniqueOpens: number
  clicks: number
  uniqueClicks: number
  forwards: number
  forwardOpens: number
  openRate: number
  clickRate: number
}
// ...
export interface MailchimpEngagementResponse {
  data: MailchimpEngagementData[]
  listId: string
  _source: string
}